fix: Use client where necessary
