+++ src/app/(layout-with-navigation)/(pages-statiques)/guide/_components/Guide.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import GuideContent from './_content/guide.mdx'
+import GuideContent from '../_content/guide.mdx'
 
 export default function Guide() {
   return <GuideContent />
