+++ src/app/(layout-with-navigation)/(pages-statiques)/nouveautes/[slug]/page.tsx
@@ -1,23 +1,24 @@
 import Route404 from '@/components/layout/404'
-import Meta from '@/components/misc/MetaOpenGraph'
 import InlineLink from '@/design-system/inputs/InlineLink'
 import Title from '@/design-system/layout/Title'
 import Markdown from '@/design-system/utils/Markdown'
 import { getFormattedDate } from '@/helpers/date/getFormattedDate'
 import { getServerTranslation } from '@/helpers/getServerTranslation'
+import { getMetadataObject } from '@/helpers/metadata/getMetadataObject'
 import { getCurrentLangInfos } from '@/locales/translation'
 import { capitaliseString } from '@/utils/capitaliseString'
-import { Metadata } from 'next'
-import { extractImageSrc } from '../_helpers/extractImage'
 import { getPath } from '../_helpers/getPath'
 import { slugifyString } from '../_helpers/slugifyString'
 import { sortReleases } from '../_helpers/sortReleases'
 import ReleaseSelect from './_components/ReleaseSelect'
 
-export const metadata: Metadata = {
-  title: 'Les nouveautés - Nos Gestes Climat',
-  description:
-    'Consultez les nouvelles fonctionnalités et dernières nouvelles de Nos Gestes Climat.',
+export function generateMetadata() {
+  return getMetadataObject({
+    title: 'Les nouveautés - Nos Gestes Climat',
+    description:
+      'Consultez les nouvelles fonctionnalités et dernières nouvelles de Nos Gestes Climat.',
+    noImage: true,
+  })
 }
 
 const removeGithubIssuesReferences = (text: string) =>
@@ -49,25 +50,8 @@ export default async function NewsPage({
   const releaseName = data[selectedReleaseIndex]?.name?.toLowerCase()
   const body = data[selectedReleaseIndex]?.body
 
-  const image = extractImageSrc(body)
-
-  const releaseDateCool = getFormattedDate(
-    new Date(data[selectedReleaseIndex].published_at),
-    currentLangInfos.abrvLocale
-  )
-
   return (
     <div className="news-page flex items-start justify-center gap-8">
-      <Meta
-        title={`${t('Nouveautés')} ${releaseDateCool} - ${capitaliseString(
-          releaseName
-        )}`}
-        image={image}
-        description={t(
-          'Consultez les dernières nouveautés de Nos Gestes Climat'
-        )}
-      />
-
       <ReleaseSelect
         releases={data}
         selectedReleaseIndex={selectedReleaseIndex}
