+++ src/app/(layout-with-navigation)/(simulation)/actions/_components/actions/_components/ActionForm.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Navigation from '@/components/form/Navigation'
 import Question from '@/components/form/Question'
 import { useForm } from '@/publicodes-state'
