+++ src/app/(layout-with-navigation)/(simulation)/actions/plus/page.tsx
@@ -1,19 +1,8 @@
-'use client'
-
-import Link from '@/components/Link'
 import Trans from '@/components/translation/Trans'
-import Card from '@/design-system/layout/Card'
-import { useLocale } from '@/hooks/useLocale'
-import { useRules } from '@/hooks/useRules'
-import { useUser } from '@/publicodes-state'
 import Image from 'next/image'
-import { utils } from 'publicodes'
 
 import { getMetadataObject } from '@/helpers/metadata/getMetadataObject'
-import { getRuleTitle } from '@/helpers/publicodes/getRuleTitle'
-import { useGetPRNumber } from '@/hooks/useGetPRNumber'
-import { NGCRule, NGCRules } from '@/publicodes-state/types'
-import { useFetchDocumentation } from '../_hooks/useFetchDocumentation'
+import ActionPlusList from './_components/ActionPlusList'
 
 export function generateMetadata() {
   return getMetadataObject({
@@ -24,29 +13,6 @@ export function generateMetadata() {
 }
 
 export default function ActionList() {
-  const locale = useLocale()
-  const { user } = useUser()
-  const { data: rules } = useRules({
-    region: user.region.code,
-    lang: locale || 'fr',
-  })
-
-  const { PRNumber } = useGetPRNumber()
-
-  const { data: documentation } = useFetchDocumentation(PRNumber)
-
-  if (!documentation) {
-    return null
-  }
-
-  const plusListe = Object.entries(rules as NGCRules)
-    .map(([dottedName, rule]) => ({ ...rule, dottedName }))
-    .map((rule) => {
-      const plus = documentation?.['actions-plus/' + rule.dottedName]
-      return { ...rule, plus }
-    })
-    .filter((r) => r.plus)
-
   return (
     <div className="mt-8">
       <h2>
@@ -68,23 +34,7 @@ export default function ActionList() {
         </em>
       </p>
 
-      <ul className="grid list-none grid-cols-1 gap-4 md:grid-cols-3">
-        {plusListe.map((rule) => (
-          <li key={rule.dottedName}>
-            <Card
-              className="h-[12rem] flex-col items-center justify-center no-underline"
-              tag={Link}
-              href={'/actions/plus/' + utils.encodeRuleName(rule.dottedName)}>
-              <div className="mb-8 text-2xl">{rule.icônes || '🎯'}</div>
-              <div className="text-center">
-                {getRuleTitle(
-                  rule as NGCRule & { dottedName: string; titre: string }
-                )}
-              </div>
-            </Card>
-          </li>
-        ))}
-      </ul>
+      <ActionPlusList />
     </div>
   )
 }
