+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/NewsletterForm.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Trans from '@/components/translation/Trans'
 import Button from '@/design-system/inputs/Button'
 import CheckboxInputGroup from '@/design-system/inputs/CheckboxInputGroup'
