+++ src/app/(layout-with-navigation)/(simulation)/groupes/invitation/page.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import AutoCanonicalTag from '@/design-system/utils/AutoCanonicalTag'
 import { getMetadataObject } from '@/helpers/metadata/getMetadataObject'
 import Invitation from './_components/Invitation'
