+++ src/app/(layout-with-navigation)/northstar/page.tsx
@@ -1,11 +1,7 @@
-'use client'
-
-import Loader from '@/design-system/layout/Loader'
-
 import Title from '@/design-system/layout/Title'
+import { getServerTranslation } from '@/helpers/getServerTranslation'
 import { getMetadataObject } from '@/helpers/metadata/getMetadataObject'
-import { useEffect, useRef, useState } from 'react'
-import { useTranslation } from 'react-i18next'
+import NorthStarIframe from './_components/NorthStarIframe'
 
 export function generateMetadata() {
   return getMetadataObject({
@@ -15,26 +11,13 @@ export function generateMetadata() {
   })
 }
 
-export default function NorthStarPage() {
-  const { t } = useTranslation()
+export default async function NorthStarPage() {
+  const { t } = await getServerTranslation()
+
   const title = t('Statistiques Northstar')
 
-  const iFrameRef = useRef(null)
-
-  const [isIFrameLoaded, setIsIFrameLoaded] = useState(false)
-
-  useEffect(() => {
-    const iframeCurrent: any = iFrameRef.current
-
-    iframeCurrent?.addEventListener('load', () => setIsIFrameLoaded(true))
-
-    return () => {
-      iframeCurrent?.removeEventListener('load', () => setIsIFrameLoaded(true))
-    }
-  }, [iFrameRef])
-
   return (
-    <div className={'ui__ fluid container'}>
+    <div>
       <Title title={title} />
 
       <p>
@@ -43,16 +26,7 @@ export default function NorthStarPage() {
         )}
       </p>
 
-      {!isIFrameLoaded && <Loader />}
-
-      <iframe
-        ref={iFrameRef}
-        id="iframe-metabase-northstar"
-        title="Statistiques Northstar Metabase"
-        src="https://metabase-ngc.osc-fr1.scalingo.io/public/dashboard/0f6974c5-1254-47b4-b6d9-6e6f22a6faf7"
-        width="100%"
-        height="1800px"
-        className="border-none"></iframe>
+      <NorthStarIframe />
     </div>
   )
 }
