+++ src/app/_components/IsClientCtxProvider.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {
   PropsWithChildren,
   createContext,
